Show the training queue below buildings.
This is kinda rudimentary. In future versions I might want to show the
progress of the current production as well. It would also be good to
show what kind of unit is being made. And it might be nice to hide this
information in a HUD instead of having it out all the time.