Switch to software renderer and disable vsync.
This seems to a be a bit faster in canvas::present. Maybe just because
you don't have to communicate with the graphics card, and we aren't
doing that much stuff that could be sped up by it. IDK. I'll do another
opt pass much later, but this helps for now.